minor refactor (backend): fixed missing references after db_engine rename
